+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Removed
+- The broken `add_injector` argument from [tanjun.clients.Client.add_component][].
 
 ## [2.6.2a1] - 2022-09-05
 ### Added